+++ CHANGELOG.md
@@ -20,6 +20,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 - Add support for glob in manifest source files.
 - `parents`: Print currently selected version.
 - `packages`: Add `--version` flag to print currently used versions.
+- `packages`: Add `--targets` flag to print targets used in the corresponding manifest.
 
 ### Changed
 - `update`: Clean up alignment of manual resolution output.